Extract tests for DatabaseImporter.
